feat(etl): recurring Delivery Schedules sync + adopt writer_desired mini-ETL
- Delivery Schedules: register import_delivery_schedules as a past-scope
  manifest step so it refreshes on Re-sync Past Months / rollover / cron
  (importer, delivery_templates model, and editorial_raw_delivery_templates
  BQ table already existed — it was seed/wizard-only, on no recurring scope).
- writer_desired: new etl/build_writer_desired.py ports the planning-hub's
  seed_writer_desired.mjs to Python; daily @writer-desired manifest step
  (current scope, after @name-mappings) publishes BQ editorial_writer_desired
  with the exact 15-col contract. Name reconciliation via the central
  editorial_name_map(writer) + v_editorial_roster — DaniQ added
  "Dan Pelberg -> Daniel Pelberg" to the Writers tab, so FALLBACK_ALIASES is
  empty (no code aliases). Verified: reproduces the live table + picks up new
  responses, ym 202509..202607.
- catalogs: add editorial_writer_desired to the schema HUB_PUBLISHED family +
  GRAIN and the lineage LINEAGE (consumers.planning_hub = getWriterDesired);
  scope the lineage drift check to the governed EH surface so a curated
  cross-hub consumer no longer false-positives. Regenerated + synced.
- fix(capacity): complete the capacity int->float widening (CapacitySummary
  fields + capacity_utilization_pct) so backend mypy is clean.

+++ backend/app/schemas.py
@@ -301,9 +301,9 @@ class CapacitySummary(BaseModel):
     pod: str
     month: int
     year: int
-    total_capacity: int
-    projected_used: int
-    actual_used: int | None = None
+    total_capacity: float
+    projected_used: float
+    actual_used: float | None = None
     utilization_pct: float
     status: str  # UNDER, OPTIMAL, WARNING, OVER
